+++ Changelog.md
@@ -1,7 +1,13 @@
 # Changelog
 
-### Version 1.19.3 - 1.0.1
- - port to fabric
+### Version 1.19.3 - 1.1
+- port to fabric
+- add new command `/mobScalings` to set/remove/show the configured mob scalings
+    - `/mobScalings set <entity_type> <scaling>` => set static scaling for the entity type
+    - `/mobScalings set <entity_type> <scaling_type> <min_scaling> <max_scaling>` => set a specific scaling for the entity type
+    - `/mobScalings remove <entity_type>` => remove the scaling for the entity type
+    - `/mobScalings show` => show all configured mob scalings
+    - `/mobScalings show <entity_type>` => show the scaling for the entity type
 
 ### Version 1.19.3 - 1.0
- - initial release
+- initial release